Correct pointing on the Oculus
Add a synthetic pointer event to allow the pointer to follow
navigation. Add a snap-to-orignal position to the world drag
constraint (which should be made configurable.) Not only is this
nice to use in general, it makes Oculus mode more predictable.
Dragging with your nose is effective.  Add a virtual pointer (which
should be made optional.)