Add discard property to program specification
Previously, the opacity of a program or binding was entirely determined by
the opacity of its diffuse texture. Without a texture, the discard program
was being merged with other object depth programs. This is the reason why
light sources were casting shadows, and also the source of some seemingly
random behavior in shadow reliability.