Water fragment shader compatibility hack
Fixes GTS 240 under Win 7. Previously, the sky was black. If the water
shading was disabled, the sky worked correctly. It's almost as if a
given sampler can work either in a dependent or strictly independent
mode, but not both at once. Somehow, the conditional guarantees that
this constraint holds, which the mix-step does not.