Experimenting with pointer handling
Stereoscopic and multi-viewport modes have two issues: First, the eye offset
isn't applied to the overlay causing the pointer to not line up. Second,
the pointer's 3D position jumps when the mouse crosses from one viewport to
another.

The current hack presumes that the first frustum in the system is always
correct... that the pointer is always "in" it even when it's not within its
viewport, and that the its eye location defines the overlay eye position.
Defing a true overlay overrides all of this.