Reimplement the framebuffer stack using the attribute stack
Previously, new viewport values were pushed onto the framebuffer stack along with the framebuffer object. This stack was initialized with the 0th framebuffer and the initial viewport size. However, if the window size changed, the stack would be left with an obsolete viewport size on the stack. This reimplementation uses the attribute stack to push the current viewport, whatever it may be.