Reinstate focus check
Turns out it's not obsolete after all because the file selector might delete
and recreate its child tree in response to a click in its child.

However, we can't simply do a focus search after handling such a click because
another button may have caused the entire dialog to be deleted.

The solution is to perform a focus search prior to every click. This leaves the
GUI without a focus after a normal click, but that's not such a big deal.