Refactor GUI focus handling YET AGAIN
Seems like the best fix of all is simply to declare the GUI focus and input
as static. This ensures that their storage is retained through the deletion
of a dialog. It also allows the focus and input to be unset in the widget
destructor if and only if needed.

It's also rational that there should be only one focus and one input even
if there are multiple dialogs.